requirements.txt
torch==2.0.0
jupyter==1.0.0
ipykernel==6.22.0
jupyterlab==3.6.3
lightning==2.0.1.post0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
scikit-learn==1.2.2
seaborn==0.12.2
shiny
pytest==7.3.1
shinylive
shinywidgets
ipyleaflet
pyodide-py
